# openapi.yaml
openapi: 3.0.0
info:
  title: Faculty Reporting System API
  version: 1.0.0
  description: API documentation for the Faculty Reporting System (backend in Express.js)
servers:
  - url: http://localhost:5003/api
    description: Local development server (client .env default)
  - url: http://localhost:5000/api
    description: Local development server (server default)
  - url: http://localhost:5001/api
    description: Local development server (alternate)
security:
  - BearerAuth: []
paths:
  /auth/register:
    post:
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
  /auth/login:
    post:
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
    RegisterRequest:
      type: object
      required: [username, email, password, firstName, lastName, role]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [student, lecturer, program_leader, principal_lecturer, faculty_manager]
        phone:
          type: string
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        is_active:
          type: boolean
...